+++ src/highlighters/mod.rs
@@ -4,6 +4,7 @@ mod key_value;
 mod keyword;
 mod number;
 mod path;
+mod process;
 mod quotes;
 mod url;
 mod uuid;
@@ -14,6 +15,7 @@ use crate::highlighters::key_value::KeyValueHighlighter;
 use crate::highlighters::keyword::KeywordHighlighter;
 use crate::highlighters::number::NumberHighlighter;
 use crate::highlighters::path::PathHighlighter;
+use crate::highlighters::process::ProcessHighlighter;
 use crate::highlighters::quotes::QuoteHighlighter;
 use crate::highlighters::url::UrlHighlighter;
 use crate::highlighters::uuid::UuidHighlighter;
@@ -54,12 +56,16 @@ impl Highlighters {
             before_fns.push(Box::new(IpHighlighter::new(&ip.segment, &ip.separator)));
         }
 
-        if let Some(key_value) = &config.groups.key_value {
-            before_fns.push(Box::new(KeyValueHighlighter::new(&key_value.key, &key_value.separator)));
+        if let Some(kv) = &config.groups.key_value {
+            before_fns.push(Box::new(KeyValueHighlighter::new(&kv.key, &kv.separator)));
         }
 
         if let Some(uuid) = &config.groups.uuid {
             before_fns.push(Box::new(UuidHighlighter::new(&uuid.segment, &uuid.separator)));
+        }
+
+        if let Some(p) = &config.groups.process {
+            before_fns.push(Box::new(ProcessHighlighter::new(&p.name, &p.separator, &p.id)));
         }
 
         before_fns